requirements.txt
asgiref==3.9.1
Django==5.2.4
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
